# openapi.yaml
openapi: 3.0.3
info:
  title: Donation Platform API
  version: 1.0.0
  description: |
    This API handles the core functionality of the Donation Management Platform.  
    It allows users to register, authenticate, and interact with donation posts.

    Donors can create and manage donation entries, while charities can browse,
    filter, and claim available donations. The system ensures proper tracking of
    quantities to prevent over-claiming and maintains clear communication between users.

    Authentication is handled using JSON Web Tokens (JWT), and access to certain
    endpoints is restricted based on user roles (donor, charity, admin).

    This documentation provides details on all available endpoints, request formats,
    and expected responses.
    TODO: Confirm exact success status codes where the documentation only says "Success response".
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
paths:
...
  /claims:
    post:
      tags: [Claims]
      summary: Claim a donation
      description: Creates a claim for part or all of a donation. Approved charity required.
      operationId: createClaim
      security: [{bearerAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/CreateClaim'}
      responses:
        "201": {$ref: '#/components/responses/ClaimCreated'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
    get:
      tags: [Claims]
      summary: Get all claims
      description: Returns all claims. Admin only.
      operationId: getAllClaims
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ClaimStatusQuery'}
      responses:
        "200":
          description: Claims returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/AdminClaim'}}
              example: [{_id: CLAIM_ID, donation: {_id: DONATION_ID, name: Rice bags, donationType: food}, charity: {_id: CHARITY_USER_ID, name: Haithem, email: haithem@gmail.com, organization: KhirCharity}, quantity: 5, status: pending, notes: KhirCharity can collect the items this week.}]
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
...
  /claims/my:
    get:
      tags: [Claims]
      summary: Get my claims
      description: Returns claims created by the logged-in charity.
      operationId: getMyClaims
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Charity claims returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/MyClaim'}}
              example: [{_id: CLAIM_ID, donation: {_id: DONATION_ID, name: Rice bags, donationType: food}, charity: CHARITY_USER_ID, quantity: 5, status: pending, notes: KhirCharity can collect the items this week.}]
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /claims/{id}:
    get:
      tags: [Claims]
      summary: Get one claim
      description: Viewable by admins, the charity that created the claim, or the donor who owns the donation.
      operationId: getClaimById
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        "200":
          description: Claim returned.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClaimDetail'}
              example: {_id: CLAIM_ID, donation: {_id: DONATION_ID, name: Rice bags, donor: DONOR_ID}, charity: {_id: CHARITY_USER_ID, name: Haithem, email: haithem@gmail.com, organization: KhirCharity}, quantity: 5, status: pending, notes: KhirCharity can collect the items this week.}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /claims/{id}/status:
    put:
      tags: [Claims]
      summary: Update claim status
      description: Updates a claim status. Admin only. Rejected claims restore quantity; returning rejected claims to pending or confirmed reserves quantity again.
      operationId: updateClaimStatus
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateClaimStatusRequest'}
            example: {status: confirmed}
      responses:
        "200":
          description: Claim status updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UpdatedClaim'}
              example: {_id: CLAIM_ID, donation: {_id: DONATION_ID, name: Rice bags, donationType: food}, charity: CHARITY_USER_ID, quantity: 5, status: confirmed, notes: KhirCharity can collect the items this week.}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /notifications:
    get:
      tags: [Notifications]
      summary: Get notifications
      description: Returns notifications for the logged-in user.
      operationId: getNotifications
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Notifications returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Notification'}}
              example: [{_id: NOTIFICATION_ID, user: USER_ID, type: donation_claimed, message: 'KhirCharity claimed 5 units from "Rice bags". 15 units remain.', isRead: false, createdAt: "2026-04-12T20:30:00.000Z", updatedAt: "2026-04-12T20:30:00.000Z"}]
        "401": {$ref: '#/components/responses/Unauthorized'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      operationId: markNotificationAsRead
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        "200":
          description: Notification marked as read.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MessageResponse'}
              example: {message: Notification marked as read}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Delete notification
      operationId: deleteNotification
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        "200":
          description: Notification deleted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MessageResponse'}
              example: {message: Notification deleted successfully}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token returned by POST /auth/login.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema: {type: string}
      example: RESOURCE_ID
...
    ClaimStatusQuery:
      name: status
      in: query
      required: false
      description: Filter claims by status.
      schema: {$ref: '#/components/schemas/ClaimStatus'}
      example: pending
...
  requestBodies:
    CreateClaim:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CreateClaimRequest'}
          example: {donationId: DONATION_ID, quantity: 5, notes: KhirCharity can collect the items this week.}
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            validation: {value: {message: Invalid request data}}
            invalidDonationType: {value: {message: Invalid donation type}}
            invalidStatus: {value: {message: Invalid claim status}}
    Unauthorized:
      description: Not authenticated.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            noToken: {value: {message: No token}}
            invalidToken: {value: {message: Invalid token}}
            tokenExpired: {value: {message: Token expired}}
            userNotFound: {value: {message: User not found}}
    Forbidden:
      description: Not authorized.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            actionNotAuthorized: {value: {message: Action not authorized}}
            accessDenied: {value: {message: Access denied}}
            charityNotVerified: {value: {message: Charity not verified yet}}
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Resource not found}
...
    RateLimit:
      description: Too many requests.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Too many requests, please try again later}
...
    ServerError:
      description: Server error.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Server error message}
    ClaimCreated:
      description: "Claim created. TODO: Confirm exact status code."
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CreateClaimResponse'}
          example:
            message: Donation claimed successfully
            transaction: {_id: CLAIM_ID, donation: DONATION_ID, charity: CHARITY_USER_ID, quantity: 5, status: pending, claimedAt: "2026-04-12T20:30:00.000Z", notes: KhirCharity can collect the items this week., createdAt: "2026-04-12T20:30:00.000Z", updatedAt: "2026-04-12T20:30:00.000Z"}
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message: {type: string, example: Error message here}
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: {type: string, example: Operation completed successfully}
...
    DonationType:
      type: string
      enum: [food, clothes, toys, electronics, books, medicines, furniture, school_supplies, hygiene, baby_items, household, other, others]
...
    ClaimStatus:
      type: string
      enum: [pending, confirmed, rejected]
    NotificationType:
      type: string
      enum: [charity_approved, charity_rejected, donation_claimed, donation_failed_insufficient_quantity]
...
    CharitySummary:
      type: object
      required: [_id, name, email, organization]
      properties:
        _id: {type: string, example: CHARITY_USER_ID}
        name: {type: string, example: Haithem}
        email: {type: string, format: email, example: haithem@gmail.com}
        organization: {type: string, example: KhirCharity}
...
    ClaimDonationSummary:
      type: object
      required: [_id, name]
      properties:
        _id: {type: string, example: DONATION_ID}
        name: {type: string, example: Rice bags}
        donationType: {$ref: '#/components/schemas/DonationType'}
        donor: {type: string, example: DONOR_ID}
    ClaimBase:
      type: object
      required: [_id, quantity, status]
      properties:
        _id: {type: string, example: CLAIM_ID}
        quantity: {type: integer, minimum: 1, example: 5}
        status: {$ref: '#/components/schemas/ClaimStatus'}
        notes: {type: string, example: KhirCharity can collect the items this week.}
...
    AdminClaim:
      allOf:
        - {$ref: '#/components/schemas/ClaimBase'}
        - type: object
          required: [donation, charity]
          properties:
            donation: {$ref: '#/components/schemas/ClaimDonationSummary'}
            charity: {$ref: '#/components/schemas/CharitySummary'}
    MyClaim:
      allOf:
        - {$ref: '#/components/schemas/ClaimBase'}
        - type: object
          required: [donation, charity]
          properties:
            donation: {$ref: '#/components/schemas/ClaimDonationSummary'}
            charity: {type: string, example: CHARITY_USER_ID}
    ClaimDetail:
      allOf:
        - {$ref: '#/components/schemas/ClaimBase'}
        - type: object
          required: [donation, charity]
          properties:
            donation: {$ref: '#/components/schemas/ClaimDonationSummary'}
            charity: {$ref: '#/components/schemas/CharitySummary'}
    UpdatedClaim:
      allOf:
        - {$ref: '#/components/schemas/ClaimBase'}
        - type: object
          required: [donation, charity]
          properties:
            donation: {$ref: '#/components/schemas/ClaimDonationSummary'}
            charity: {type: string, example: CHARITY_USER_ID}
    CreateClaimRequest:
      type: object
      required: [donationId, quantity]
      properties:
        donationId: {type: string, example: DONATION_ID}
        quantity: {type: integer, minimum: 1, example: 5}
        notes: {type: string, example: KhirCharity can collect the items this week.}
    ClaimTransaction:
      allOf:
        - {$ref: '#/components/schemas/ClaimBase'}
        - type: object
          required: [donation, charity]
          properties:
            donation: {type: string, example: DONATION_ID}
            charity: {type: string, example: CHARITY_USER_ID}
            claimedAt: {type: string, format: date-time, example: "2026-04-12T20:30:00.000Z"}
            createdAt: {type: string, format: date-time, example: "2026-04-12T20:30:00.000Z"}
            updatedAt: {type: string, format: date-time, example: "2026-04-12T20:30:00.000Z"}
    CreateClaimResponse:
      type: object
      required: [message, transaction]
      properties:
        message: {type: string, example: Donation claimed successfully}
        transaction: {$ref: '#/components/schemas/ClaimTransaction'}
    UpdateClaimStatusRequest:
      type: object
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/ClaimStatus'}
    Notification:
      type: object
      required: [_id, user, type, message, isRead, createdAt, updatedAt]
      properties:
        _id: {type: string, example: NOTIFICATION_ID}
        user: {type: string, example: USER_ID}
        type: {$ref: '#/components/schemas/NotificationType'}
        message: {type: string, example: 'KhirCharity claimed 5 units from "Rice bags". 15 units remain.'}
        isRead: {type: boolean, example: false}
        createdAt: {type: string, format: date-time, example: "2026-04-12T20:30:00.000Z"}
        updatedAt: {type: string, format: date-time, example: "2026-04-12T20:30:00.000Z"}